+++ src/model/Attack.java
@@ -2,15 +2,15 @@ package model;
 
 public class Attack {
 
-	public byte range;
+	public int range;
 	public AttackType attackType;
 	public double meleeMultiplier;
 	public double rangeMultiplier;
 	public boolean chain;
 	public boolean push;
 	
-	public Attack(byte range, AttackType attackType,
-			short damage, double meleeMultiplier, double rangeMultiplier,
+	public Attack(int range, AttackType attackType,
+			int damage, double meleeMultiplier, double rangeMultiplier,
 			boolean chain, boolean push) {
 		super();
 		this.range = range;
