+++ src/model/Heal.java
@@ -2,11 +2,11 @@ package model;
 
 public class Heal {
 
-	public byte range;
-	public short heal;
-	public short revive;
+	public int range;
+	public int heal;
+	public int revive;
 	
-	public Heal(byte range, short heal, short revive) {
+	public Heal(int range, int heal, int revive) {
 		super();
 		this.range = range;
 		this.heal = heal;
